requirements.txt
matplotlib~=3.9.0
pillow~=10.3.0
numpy~=1.26.4
scikit-learn~=1.5.0